Fake Database progress
Included probably temporary variable "id" for item.  I expect it to be a
primary key for the real database later on.
Squashed all the information in a map into a single string for the
database, and wrote a method to extract the information as well.
Currently waiting for item instances for these methods to become
functional.