Fake Database Initial Data class is complete
it properly gathers all possible data, while leaving the unknown data to
be filled later.
example: Actor inventory contains items, but the items list (per
account) and actor aren't connected yet.  The database will only
remember the item ID for the inventory, leaving the rest of the item
constructor blank.  When the game is loaded (and the account ID is
chosen through login credentials), the item list will be built, and code
will run through the inventory and fill in the blank data.