Load Game works in Fake Database
The database code runs through all the lists extracted from the tables.
Given a username and password, it will find the account ID, and all the
data associated with it. It will save this data to a given Game model
class.

Method was also added in GameEngine to create a new instance of an item
given an item ID.