Fake database now saves player and Creatures.
fake database class now creates account and immediately saves it to the
CSV when creating a new game.
Writing the accounts table in the database no longer occurs when saving
the game.
Provide accountId to the WriteCSV method for now on.  Should be
obtainable from make classes, map, item, etc.  Keeping the original
instance of the fake database is no longer necessary to save the game.

My last commit was denied and all the staged changes disappeared.  In
case the commit message is lost:
	GameEngine was given a method to create a list of default items for a
new game.
	Player and creature can be loaded from database