empty game log, view log command
Added a new command to go back to the game log from location or map.

Added a blank space to the game logs when creating a game.  This is just
so the player can view their log even when it's empty.  Previously, the
box wouldn't even show up if they entered the "view log" command.